Feature/sleep mode (#6)
* Modified disconnecting behavior, such that the player is no longer removed from all servers. Instead, the player is marked as 'Asleep', which puts them in a mode which does not affect voting. The sleeping players vote can still be shown, but it is not necessary that the player votes to show all votes.
When the player rejoins or reconnects, the player is awaken again, and behaves like a normal player.

* Indicate sleeping players in UI.

* Fixed issue, where sleeping player did not keep their player when rejoining, as the player connection id rotated.
This has been fixed by introducing a recovery id, which it kept in local storage. Whenever the session is refreshed, or connection is lost, the player will be recovered upon joining the server, keeping their name and vote.