Feature/12/fix 404 (#16)
Fixes #12 by doing a preliminary exists-check on the server, to ensure it is joinable. If the server does not exist, the user will be presented with a 404 page.

This is resolution B from the issue. I found this was a better fix, as the user will know whether the server exists before attempting to join, which could otherwise cause frustration, as the user would have inputted a user name before being presented with this information.