remotes/origin/bugfix/fix-flaky-tests (#27)
Fixed flaky tests, which was caused by the OnSessionUpdate being vulnerable to receiving the session state from previous updates. It cannot guarantee the order which responses are being sent back. To fix this, the test must ensure that the session updated is the right one. This is done by ensuring all previous commands are waiting for the session update that belongs to that specific action.

Additional minor fixes:
* Introduced shared default timeout provider, to keep the tests DRY.